fix memcheck
