+++ src/arithmetic/boolean_funcs/boolean_funcs.c
@@ -9,6 +9,8 @@
 #include "../func_desc.h"
 #include "../../util/arr.h"
 #include "../../query_ctx.h"
+#include "../../datatypes/map.h"
+#include "../../datatypes/array.h"
 
 #define CONTAINS_NULL 2 // Macro used for efficiently evaluating 3-valued truth table
 
@@ -196,6 +198,26 @@ SIValue AR_TO_BOOLEAN(SIValue *argv, int argc, void *private_data) {
 	}
 }
 
+SIValue AR_ISEMPTY(SIValue *argv, int argc, void *private_data) {
+	ASSERT(argc == 1);
+	switch(SI_TYPE(argv[0])) {
+		case T_NULL:
+			return SI_NullVal();
+		case T_ARRAY:
+			if(SIArray_Length(argv[0]) == 0) return SI_BoolVal(true);
+			break;
+		case T_MAP:
+			if(array_len(argv[0].map) == 0) return SI_BoolVal(true);
+			break;
+		case T_STRING:
+			if(strlen(argv[0].stringval) == 0) return SI_BoolVal(true);
+			break;
+		default:
+			ASSERT(false);
+	}
+	return SI_BoolVal(false);
+}
+
 void Register_BooleanFuncs() {
 	SIType *types;
 	SIType ret_type = T_BOOL | T_NULL;
@@ -278,5 +300,11 @@ void Register_BooleanFuncs() {
 	types = array_new(SIType, 1);
 	array_append(types, SI_ALL);
 	func_desc = AR_FuncDescNew("toBooleanOrNull", AR_TO_BOOLEAN, 1, 1, types, ret_type, false, true);
+	AR_RegFunc(func_desc);
+
+	types = array_new(SIType, 1);
+	array_append(types, T_ARRAY | T_MAP | T_NULL | T_STRING);
+	ret_type = T_BOOL | T_NULL;
+	func_desc = AR_FuncDescNew("isempty", AR_ISEMPTY, 1, 1, types, ret_type, false, true);
 	AR_RegFunc(func_desc);
 }
